[DELIVERS #135877245]
    - Moves every thest to its separate file and change database names
      to be unique for each test/file. This makes the tests independent
and not crap out when run concurrently.
