+++ test/_config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  usename: 'root'
+, password: 'password'
+}
